SEMANA07/dia2/ecommerce/api/serializers.py
from rest_framework import serializers

from web.models import Categoria,Producto

class CategoriaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Categoria
        fields = '__all__'