SEMANA07/dia2/ecommerce/requirements.txt
django==3.2
mysqlclient
#psycopg2
Pillow
django-ckeditor
django-admin-tailwind
django-paypal
#gunicorn
djangorestframework